requirements.txt
numpy==1.24.2
opencv-python-headless
Pillow==9.5.0
PyYAML==6.0
streamlit==1.19.0
